THE_WORLD/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
GeoAlchemy2==0.14.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.1
